move device dao sembast impl
